--- a/thirdparty/DiskANN/src/pq_flash_index.cpp
+++ b/thirdparty/DiskANN/src/pq_flash_index.cpp
@@ -54,6 +54,7 @@
   ((((_u64) (id)) % nvecs_per_sector) * data_dim * sizeof(float))
 
 namespace {
+  constexpr size_t kReadBatchSize = 32;
   constexpr _u64 kRefineBeamWidthFactor = 2;
   constexpr _u64 kBruteForceTopkRefineExpansionFactor = 2;
   auto           calcFilterThreshold = [](const auto topk) -> const float {
@@ -1527,31 +1528,57 @@ namespace diskann {
   }
 
   template<typename T>
-  void PQFlashIndex<T>::get_vector_by_sector(const _u64 sector_offset,
-                                             const std::vector<_u64> &ids_idx,
-                                             const int64_t           *ids,
-                                             T *output_data) {
+  void PQFlashIndex<T>::get_vector_by_ids(const int64_t *ids, const int64_t n, T *output_data) {
+    size_t batch_size = kReadBatchSize;
+    if (long_node) {
+      auto min_size = kReadBatchSize / nsectors_per_node;
+      batch_size = (min_size == 0) ? MAX_N_SECTOR_READS : min_size;
+      if (0 == batch_size) {
+        LOG(ERROR) << "Vector too large, exceeding max number of sector reads";
+        return ;
+      }
+    }
     ThreadData<T> data = this->thread_data.pop();
     while (data.scratch.sector_scratch == nullptr) {
       this->thread_data.wait_for_push_notify();
       data = this->thread_data.pop();
     }
-    auto                     ctx = this->reader->get_ctx();
-    char                    *sector_scratch = data.scratch.sector_scratch;
-    _u64                    &sector_scratch_idx = data.scratch.sector_idx;
+    char *sector_scratch = data.scratch.sector_scratch;
     std::vector<AlignedRead> frontier_read_reqs;
+    frontier_read_reqs.reserve(batch_size);
 
-    char *sector_buf = sector_scratch + sector_scratch_idx * read_len_for_node;
-    frontier_read_reqs.emplace_back(sector_offset, read_len_for_node,
+    auto sectors_to_visit = get_sectors_layout_and_write_data_from_cache(ids, n, output_data);
+    if (sectors_to_visit.size() == 0) {
+      return;
+    }
+
+    std::vector<_u64> sector_offsets;
+    sector_offsets.reserve(sectors_to_visit.size());
+    for (const auto &it : sectors_to_visit) {
+      sector_offsets.emplace_back(it.first);
+    }
+
+    auto ctx = this->reader->get_ctx();
+    const auto sector_num = sector_offsets.size();
+    const _u64 num_blocks = DIV_ROUND_UP(sector_num, batch_size);
+    for (_u64 i = 0; i < num_blocks; ++i) {
+      _u64 start_idx = i * batch_size;
+      _u64 idx_len = std::min(batch_size, sector_num - start_idx);
+      frontier_read_reqs.clear();
+      for (_u64 j = 0; j < idx_len; ++j) {
+        char *sector_buf = sector_scratch + j * read_len_for_node;
+        frontier_read_reqs.emplace_back(sector_offsets[start_idx + j], read_len_for_node,
                                     sector_buf);
-#ifdef USE_BING_INFRA
-    reader->read(frontier_read_reqs, ctx, true);  // async reader windows.
-#else
-    reader->read(frontier_read_reqs, ctx);  // synchronous IO linux
-#endif
-    for (const auto idx : ids_idx) {
-      char *node_buf = get_offset_to_node(sector_buf, ids[idx]);
-      copy_vec_base_data(output_data, idx, node_buf);
+      }
+      reader->read(frontier_read_reqs, ctx);
+      for (const auto& req : frontier_read_reqs) {
+        auto offset = req.offset;
+        char* sector_buf = static_cast<char*>(req.buf);
+        for (auto idx : sectors_to_visit[offset]) {
+          char* node_buf = get_offset_to_node(sector_buf, ids[idx]);
+          copy_vec_base_data(output_data, idx, node_buf);
+        }
+      }
     }
     this->thread_data.push(data);
     this->thread_data.push_notify_all();