@PwzXxm 2023-06-15
Vec same as before, test script without sleep(300ms)
Get 50 ids, measure the time of 200 iterations, DiskANN cache off
Initial runs are ignored

  | Before (s) | After (s)
-- | -- | --
Concurrency 1 | 0.151025 | 0.1153
Concurrency 4 | 0.431405 | 0.292123
Concurrency 20 | 1.86118 | 0.962817

Decreased around 23% of total time and 50% when concurrency was high. It should now perform much better when concurrency is high or the number of requested IDs is large. 

The fluctuations of I/O latency are considered normal to NVMe SSD.